requirements.txt
django==4.1.7
django-environ==0.10.0
Pillow==9.4.0
python-slugify==8.0.1



